+++ CRA_TeamProject_SSD/CRA_TeamProject_SSD.vcxproj
@@ -128,6 +128,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="CommandFactory.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="TestApplication.cpp" />
     <ClCompile Include="TestShell.cpp" />
