+++ CRA_TeamProject_SSD/CRA_TeamProject_SSD.vcxproj.filters
@@ -36,6 +36,9 @@
     <ClCompile Include="WriteCommand.cpp">
       <Filter>소스 파일\Command</Filter>
     </ClCompile>
+    <ClCompile Include="CommandFactory.cpp">
+      <Filter>소스 파일\Command</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ISSD.h">
