+++ CRA_TeamProject_SSD/TestShell.cpp
@@ -7,7 +7,6 @@
 #include <vector>
 #include <stdexcept>
 
-#include "CommandFactory.h"
 
 TestShell::TestShell(ISSD* ssd) :
     _app{new TestApplication()},
